ebneuro: Don't use constructor for the preset configuration
The "default preset" was defined via the object struct which would
mean that the user would have to edit those values to change the preset
before calling the upload method. Instead let's set those in the
preset creation method itself so there is less insentive in touching
the internal state.

Signed-off-by: Nikita Travkin <nikita@trvn.ru>